tests/wpt/tests/html/canvas/element/layers/2d.layer.putImageData.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.layer.putImageData</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.layer.putImageData</h1>
<p class="desc">Check that calling putImageData in a layer throws an exception.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Check that calling putImageData in a layer throws an exception.");
_addTest(function(canvas, ctx) {

  const canvas2 = new OffscreenCanvas(100, 50);
  const ctx2 = canvas2.getContext('2d')
  const data = ctx2.getImageData(0, 0, 1, 1);
  // `putImageData` shouldn't throw on it's own.
  ctx.putImageData(data, 0, 0);
  // Make sure the exception isn't caused by calling the function twice.
  ctx.putImageData(data, 0, 0);
  // Calling again inside a layer should throw.
  ctx.beginLayer();
  assert_throws_dom("InvalidStateError", () => ctx.putImageData(data, 0, 0));

});
</script>
